Fix unaligned access exception in gptp, add SW hook (#22)
* Fix unaligned access exception

* Latest gptp snapshot

* Clean prop delay workaround for lowrisc_eth.

* Remove prop_delay workaround.
Timestamp issue fixed in lowrisc eth driver.

+++ subsys/net/l2/ethernet/gptp/gptp_data_set.h
@@ -277,7 +277,7 @@ struct gptp_default_ds {
 	bool gm_capable;
 
 	/** Number of ports of the time-aware system. */
-	uint8_t nb_ports;
+	uint32_t nb_ports;
 
 	/** Primary priority of the time-aware system. */
 	uint8_t priority1;
